+++ example/index.ts
@@ -1,3 +1,3 @@
 import { yo } from '../lib/index';
 
-yo(process.argv[2] || 'yoo-hoo', { spacing: 1, color: 'random' });
+yo(process.argv[2] || 'yoo-hoo', { spacing: 1, color: 'rainbow', paddingStart: 5 });